pyproject.toml
[project]
name = "textract-tiquu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler>=3.11.0",
    "bytez>=1.0.4",
    "chromadb>=1.0.15",
    "docling>=2.41.0",
    "docling-core[chunking,chunking-openai]>=2.43.0",
    "dspy>=2.6.27",
    "dspy-ai>=2.6.27",
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "huggingface-hub>=0.33.4",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-docling>=1.0.0",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.8",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.28",
    "litellm>=1.75.3",
    "marker-pdf>=1.8.3",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "openai>=1.97.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pip>=25.1.1",
    "pydantic>=2.11.7",
    "pypdf>=5.8.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "ragas>=0.3.0",
    "rich>=14.0.0",
    "semantic-chunkers>=0.1.1",
    "sentence-transformers>=5.0.0",
    "setuptools>=80.9.0",
    "spacy>=3.8.7",
    "spacy-llm>=0.7.3",
    "torch>=2.7.1",
    "transformers>=4.51.3",
    "unstructured-inference>=1.0.5",
    "unstructured-pytesseract>=0.3.15",
    "wheel>=0.45.1",
    "wtpsplit>=2.1.6",
]
